auth_service/requirements.txt
# ================================================================
# Auth Service - Python Dependencies
# ================================================================

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database - AsyncPG only (no ORM)
asyncpg==0.29.0

# Security & Cryptography
bcrypt==4.1.1
PyJWT==2.10.1

# HTTP Client (for inter-service calls)
aiohttp==3.9.1

# Environment Management
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
